+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Support for correct UTF8 orderby for post_title and term name (äöå) hooks now always in, before in some rare occasion
 
 ## [2.1.3-2.1.4]
 ### Added